feat: adding department field in user update and removing from equipment update
